favorite: use more appropriate icon than 'help about'
In selecting this, these factors were considered:

* use symbolic icon as elsewhere and Elementary style guidelines
* use action for icons on buttons

The current icon appears to have been selected for its visual
representation in the Elementary icon theme
(apparently "help about" is a star)
which is the wrong way around.

Using appropriate icon name is needed to work well with other
themes, or if Elementary changes the icon in the future.